--- a/lib/types/results-wrapper.js
+++ b/lib/types/results-wrapper.js
@@ -12,78 +12,75 @@ const Row = require("./row");
 const Tuple = require("./tuple");
 
 /**
- * Checks the type of value wrapper object and gets it from the underlying value
+ * Maps the value returned from the Rust driver into expected JS object.
+ *
+ * This function accepts the following kinds of arguments:
+ *   1. `Value` ->  When can correctly handle the value based on the type;
+ *   2. `[Type, Value]` ->  In case of ambiguity, rust side provides tuple of (CqlType, Value);
+ *   3. `Array` -> Tuple in JS is represented as an array so we need to distinguish between the previous case
+ *      and a case where the array is provided as value. We determine it based on the type of the first value.
+ *
+ * See rust.CqlValueWrapper, for more information
  * @param {rust.CqlValueWrapper} field
  * @returns {any}
  */
 function getCqlObject(field) {
-    if (field == null) return null;
-    let type = field.getType();
-    let res, fields;
-    switch (type) {
-        case rust.CqlType.Ascii:
-            return field.getAscii();
-        case rust.CqlType.BigInt:
-            return bigintToLong(field.getBigint());
-        case rust.CqlType.Blob:
-            return field.getBlob();
-        case rust.CqlType.Boolean:
-            return field.getBoolean();
-        case rust.CqlType.Counter:
-            return field.getCounter();
-        case rust.CqlType.Date:
-            return LocalDate.fromRust(field.getLocalDate());
-        case rust.CqlType.Double:
-            return field.getDouble();
-        case rust.CqlType.Duration:
-            return Duration.fromRust(field.getDuration());
-        case rust.CqlType.Empty:
+    switch (true) {
+        case field === null:
             return null;
-        case rust.CqlType.Float:
-            return field.getFloat();
-        case rust.CqlType.Int:
-            return field.getInt();
-        case rust.CqlType.Set:
-            res = [];
-            fields = field.getSet();
-            for (let i = 0; i < fields.length; i++)
-                res.push(getCqlObject(fields[i]));
-            return res;
-        case rust.CqlType.Text:
-            return field.getText();
-        case rust.CqlType.Timestamp:
-            // Currently only values inside Date safe range are supported
-            // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#the_epoch_timestamps_and_invalid_date
-            // The same problem exists in Datastax driver. This probably should be fixed at some point,
-            // but it's unlikely someone will need timestamps almost 300.000 years into the future.
-            return new Date(Number(field.getTimestamp()));
-        case rust.CqlType.List:
-            fields = field.getList();
-            return fields.map((value) => getCqlObject(value));
-        case rust.CqlType.Map:
-            fields = field.getMap();
-            res = {};
-            for (let i = 0; i < fields.length; i++)
-                res[getCqlObject(fields[i][0])] = getCqlObject(fields[i][1]);
-            return res;
-        case rust.CqlType.SmallInt:
-            return field.getSmallInt();
-        case rust.CqlType.TinyInt:
-            return field.getTinyInt();
-        case rust.CqlType.Uuid:
-            return Uuid.fromRust(field.getUuid());
-        case rust.CqlType.Timeuuid:
-            return TimeUuid.fromRust(field.getTimeUuid());
-        case rust.CqlType.Time:
-            return LocalTime.fromRust(field.getLocalTime());
-        case rust.CqlType.Inet:
-            return InetAddress.fromRust(field.getInet());
-        case rust.CqlType.Tuple:
-            return Tuple.fromRust(field.getTuple());
-        default:
-            // Temporarily returning string representation of the field
-            return field.stringify();
+        case field instanceof rust.LocalDateWrapper:
+            return LocalDate.fromRust(field);
+        case field instanceof rust.DurationWrapper:
+            return Duration.fromRust(field);
+        case field instanceof rust.TimeUuidWrapper:
+            return TimeUuid.fromRust(field);
+        case field instanceof rust.UuidWrapper:
+            return Uuid.fromRust(field);
+        case field instanceof rust.InetAddressWrapper:
+            return InetAddress.fromRust(field);
+        case field instanceof rust.LocalTimeWrapper:
+            return LocalTime.fromRust(field);
     }
+
+    if (Array.isArray(field)) {
+        // Check if the returned value is in (Type, Value) format
+        if (!(field.length === 2 && field[0] instanceof rust.CqlTypeClass)) {
+            // If not, returned value represents array. We need to map each element of the array
+            return field.map((value) => getCqlObject(value));
+        }
+
+        let res;
+        let value = field[1];
+        switch (field[0].typ) {
+            case rust.CqlType.BigInt:
+                return bigintToLong(value);
+            case rust.CqlType.Counter:
+                return value;
+            case rust.CqlType.Timestamp:
+                // Currently only values inside Date safe range are supported
+                // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#the_epoch_timestamps_and_invalid_date
+                // The same problem exists in Datastax driver. This probably should be fixed at some point,
+                // but it's unlikely someone will need timestamps almost 300.000 years into the future.
+                return new Date(Number(value));
+            case rust.CqlType.Map:
+                res = {};
+                for (const keyValuePair of value) {
+                    res[getCqlObject(keyValuePair[0])] = getCqlObject(
+                        keyValuePair[1],
+                    );
+                }
+                return res;
+            case rust.CqlType.Tuple:
+                return Tuple.fromArray(
+                    value.map((e) =>
+                        e === null ? undefined : getCqlObject(e),
+                    ),
+                );
+            default:
+                throw new Error("Unexpected type");
+        }
+    }
+    return field;
 }
 
 /**