--- a/lib/types/tuple.js
+++ b/lib/types/tuple.js
@@ -1,7 +1,5 @@
 "use strict";
 
-const resultWrapper = require("./results-wrapper");
-
 /** @module types */
 
 /**
@@ -104,19 +102,6 @@ class Tuple {
     values() {
         // Clone the elements
         return this.elements.slice(0);
-    }
-
-    /**
-     * Get tuple from rust object. Not intended to be exposed in the API.
-     * @param {Array<resultWrapper.CqlValueWrapper | null>} rustTuple
-     * @returns {Tuple}
-     * @package
-     */
-    static fromRust(rustTuple) {
-        let elements = rustTuple.map((e) =>
-            e === null ? undefined : resultWrapper.getCqlObject(e),
-        );
-        return new Tuple(...elements);
     }
 }