--- a/src/result.rs
+++ b/src/result.rs
@@ -4,15 +4,12 @@ use crate::{
     types::{
         local_date::LocalDateWrapper,
         time_uuid::TimeUuidWrapper,
-        type_wrappers::{ComplexType, CqlType},
+        type_wrappers::{ComplexType, CqlType, CqlTypeClass},
         uuid::UuidWrapper,
     },
     utils::err_to_napi,
 };
-use napi::{
-    bindgen_prelude::{BigInt, Buffer, ToNapiValue},
-    Error, Status,
-};
+use napi::bindgen_prelude::{BigInt, Buffer, ToNapiValue};
 use scylla::{
     cluster::metadata::{CollectionType, NativeType},
     errors::IntoRowsResultError,
@@ -24,7 +21,6 @@ use scylla::{
 use crate::types::duration::DurationWrapper;
 use crate::types::inet::InetAddressWrapper;
 use crate::types::local_time::LocalTimeWrapper;
-use crate::utils::js_error;
 
 enum QueryResultVariant {
     EmptyResult(QueryResult),
@@ -56,10 +52,16 @@ pub struct PagedRowWrapper {
 
 /// Wrapper for a single CQL value
 ///
-/// CqlValueWrapper exposes functions get____ for each CQL type,
-/// which extract value from the object,
-/// assuming this wrapper stores a value of the requested CQL type.
-#[napi]
+/// When returned from NAPI-RS, returns either just a inner value or tuple of elements: ``(cqlType, value)``
+///
+/// Plain value is returned, when we know how to treat the value, based on the value type.
+/// The tuple is returned when there is ambiguity how the given value
+/// should be handled, based on its type.
+///
+/// For example if we return String, no matter if it's CqlAscii or CqlText,
+/// we treat it the same, and don't need to specify CqlType.
+/// On the other hand, CqlBigInt and CqlTimestamp are returned from rust layer
+/// as BigInt, but are provided to driver's user as BigInt and Date, respectively.
 pub struct CqlValueWrapper {
     pub(crate) inner: CqlValue,
 }
@@ -210,246 +212,128 @@ impl PagedRowWrapper {
     }
 }
 
-#[napi]
-impl CqlValueWrapper {
-    #[napi]
-    /// This uses rust Debug to return string representation of underlying value
-    pub fn stringify(&self) -> String {
-        format!("{:?}", self.inner)
+/// Converts value into tuple of elements:
+/// ``(typ, value)``
+/// for a given CQL Value.
+///
+/// Intended to be used when there is ambiguity how the given value
+/// should be provided, based on its type.
+///
+/// # Safety
+///
+/// Valid pointer to napi env must be provided
+unsafe fn add_type_to_napi_obj(
+    env: napi::sys::napi_env,
+    value: napi::Result<napi::sys::napi_value>,
+    typ: CqlType,
+) -> napi::Result<napi::sys::napi_value> {
+    // Caller of this function ensures a valid pointer to napi env is provided
+    unsafe {
+        // We need to use CqlTypeClass for JS to recognize the value if of enum type
+        Vec::to_napi_value(
+            env,
+            vec![
+                CqlTypeClass::to_napi_value(env, CqlTypeClass { typ }),
+                value,
+            ],
+        )
     }
+}
 
-    #[napi]
-    /// Get type of CQL value that is stored in this object
-    pub fn get_type(&self) -> CqlType {
-        match &self.inner {
-            CqlValue::Ascii(_) => CqlType::Ascii,
-            CqlValue::BigInt(_) => CqlType::BigInt,
-            CqlValue::Boolean(_) => CqlType::Boolean,
-            CqlValue::Blob(_) => CqlType::Blob,
-            CqlValue::Counter(_) => CqlType::Counter,
-            CqlValue::Decimal(_) => CqlType::Decimal, // NOI
-            CqlValue::Date(_) => CqlType::Date,
-            CqlValue::Double(_) => CqlType::Double,
-            CqlValue::Duration(_) => CqlType::Duration,
-            CqlValue::Empty => CqlType::Empty,
-            CqlValue::Float(_) => CqlType::Float,
-            CqlValue::Int(_) => CqlType::Int,
-            CqlValue::Text(_) => CqlType::Text,
-            CqlValue::Timestamp(_) => CqlType::Timestamp,
-            CqlValue::Inet(_) => CqlType::Inet,
-            CqlValue::List(_) => CqlType::List,
-            CqlValue::Map(_) => CqlType::Map,
-            CqlValue::Set(_) => CqlType::Set,
-            CqlValue::UserDefinedType { .. } => CqlType::UserDefinedType, // NOI
-            CqlValue::SmallInt(_) => CqlType::SmallInt,
-            CqlValue::TinyInt(_) => CqlType::TinyInt,
-            CqlValue::Time(_) => CqlType::Time,
-            CqlValue::Timeuuid(_) => CqlType::Timeuuid,
-            CqlValue::Tuple(_) => CqlType::Tuple,
-            CqlValue::Uuid(_) => CqlType::Uuid,
-            CqlValue::Varint(_) => CqlType::Varint, // NOI
+impl ToNapiValue for CqlValueWrapper {
+    unsafe fn to_napi_value(
+        env: napi::sys::napi_env,
+        value: Self,
+    ) -> napi::Result<napi::sys::napi_value> {
+        match value.inner {
+            CqlValue::Ascii(val) => String::to_napi_value(env, val),
+            CqlValue::Boolean(val) => bool::to_napi_value(env, val),
+            CqlValue::Blob(val) => Buffer::to_napi_value(env, val.into()),
+            CqlValue::Counter(val) => add_type_to_napi_obj(
+                env,
+                BigInt::to_napi_value(env, val.0.into()),
+                CqlType::Counter,
+            ),
+            CqlValue::Decimal(_) => todo!(),
+            CqlValue::Date(val) => {
+                LocalDateWrapper::to_napi_value(env, LocalDateWrapper::from_cql_date(val))
+            }
+            CqlValue::Double(val) => f64::to_napi_value(env, val),
+            CqlValue::Duration(val) => {
+                DurationWrapper::to_napi_value(env, DurationWrapper::from_cql_duration(val))
+            }
+            CqlValue::Empty => todo!(),
+            CqlValue::Float(val) => f32::to_napi_value(env, val),
+            CqlValue::Int(val) => i32::to_napi_value(env, val),
+            CqlValue::BigInt(val) => {
+                add_type_to_napi_obj(env, BigInt::to_napi_value(env, val.into()), CqlType::BigInt)
+            }
+            CqlValue::Text(val) => String::to_napi_value(env, val),
+            CqlValue::Timestamp(val) => add_type_to_napi_obj(
+                env,
+                BigInt::to_napi_value(env, val.0.into()),
+                CqlType::Timestamp,
+            ),
+            CqlValue::Inet(val) => {
+                InetAddressWrapper::to_napi_value(env, InetAddressWrapper::from_ip_addr(val))
+            }
+            CqlValue::List(val) => Vec::to_napi_value(
+                env,
+                val.into_iter()
+                    .map(|e| CqlValueWrapper::to_napi_value(env, CqlValueWrapper { inner: e }))
+                    .collect(),
+            ),
+            CqlValue::Map(val) => add_type_to_napi_obj(
+                env,
+                Vec::to_napi_value(
+                    env,
+                    val.into_iter()
+                        .map(|e: (CqlValue, CqlValue)| {
+                            vec![
+                                CqlValueWrapper { inner: e.0 },
+                                CqlValueWrapper { inner: e.1 },
+                            ]
+                        })
+                        .collect(),
+                ),
+                CqlType::Map,
+            ),
+            CqlValue::Set(val) => Vec::to_napi_value(
+                env,
+                val.into_iter()
+                    .map(|e| CqlValueWrapper::to_napi_value(env, CqlValueWrapper { inner: e }))
+                    .collect(),
+            ),
+            CqlValue::UserDefinedType { .. } => todo!(),
+            CqlValue::SmallInt(val) => i16::to_napi_value(env, val),
+            CqlValue::TinyInt(val) => i8::to_napi_value(env, val),
+            CqlValue::Time(val) => {
+                LocalTimeWrapper::to_napi_value(env, LocalTimeWrapper::from_cql_time(val))
+            }
+            CqlValue::Timeuuid(val) => {
+                TimeUuidWrapper::to_napi_value(env, TimeUuidWrapper::from_cql_time_uuid(val))
+            }
+            CqlValue::Tuple(val) => add_type_to_napi_obj(
+                env,
+                Vec::to_napi_value(
+                    env,
+                    val.into_iter()
+                        .map(|v| {
+                            v.map(|e| {
+                                CqlValueWrapper::to_napi_value(
+                                    env,
+                                    CqlValueWrapper { inner: e.clone() },
+                                )
+                            })
+                        })
+                        .collect(),
+                ),
+                CqlType::Tuple,
+            ),
+
+            CqlValue::Uuid(val) => UuidWrapper::to_napi_value(env, UuidWrapper::from_cql_uuid(val)),
+            CqlValue::Varint(_) => todo!(),
             other => unimplemented!("Missing implementation for CQL value {:?}", other),
-        }
-    }
-
-    fn generic_error(expected_type: &str) -> Error<Status> {
-        js_error(format!(
-            "Could not get value of type {} from CqlValueWrapper",
-            expected_type
-        ))
-    }
-
-    #[napi]
-    pub fn get_ascii(&self) -> napi::Result<String> {
-        match self.inner.as_ascii() {
-            Some(r) => Ok(r.clone()),
-            None => Err(Self::generic_error("ascii")),
-        }
-    }
-
-    #[napi]
-    pub fn get_bigint(&self) -> napi::Result<BigInt> {
-        match self.inner.as_bigint() {
-            Some(r) => Ok(BigInt::from(r)),
-            None => Err(Self::generic_error("bigint")),
-        }
-    }
-
-    #[napi]
-    pub fn get_boolean(&self) -> napi::Result<bool> {
-        match self.inner.as_boolean() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("boolean")),
-        }
-    }
-
-    #[napi]
-    pub fn get_blob(&self) -> napi::Result<Buffer> {
-        match self.inner.as_blob() {
-            Some(r) => Ok(r.clone().into()),
-            None => Err(Self::generic_error("blob")),
-        }
-    }
-
-    #[napi]
-    pub fn get_counter(&self) -> napi::Result<BigInt> {
-        match self.inner.as_counter() {
-            Some(r) => Ok(r.0.into()),
-            None => Err(Self::generic_error("counter")),
-        }
-    }
-
-    #[napi]
-    pub fn get_local_date(&self) -> napi::Result<LocalDateWrapper> {
-        match self.inner.as_cql_date() {
-            Some(r) => Ok(LocalDateWrapper::from_cql_date(r)),
-            None => Err(Self::generic_error("local_date")),
-        }
-    }
-
-    #[napi]
-    pub fn get_double(&self) -> napi::Result<f64> {
-        match self.inner.as_double() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("double")),
-        }
-    }
-
-    #[napi]
-    pub fn get_duration(&self) -> napi::Result<DurationWrapper> {
-        match self.inner.as_cql_duration() {
-            Some(r) => Ok(DurationWrapper::from_cql_duration(r)),
-            None => Err(Self::generic_error("duration")),
-        }
-    }
-    #[napi]
-    pub fn get_float(&self) -> napi::Result<f32> {
-        match self.inner.as_float() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("float")),
-        }
-    }
-
-    #[napi]
-    pub fn get_int(&self) -> napi::Result<i32> {
-        match self.inner.as_int() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("int")),
-        }
-    }
-
-    #[napi]
-    pub fn get_text(&self) -> napi::Result<String> {
-        match self.inner.as_text() {
-            Some(r) => Ok(r.clone()),
-            None => Err(Self::generic_error("text")),
-        }
-    }
-
-    #[napi]
-    pub fn get_timestamp(&self) -> napi::Result<BigInt> {
-        match self.inner.as_cql_timestamp() {
-            Some(r) => Ok(r.0.into()),
-            None => Err(Self::generic_error("text")),
-        }
-    }
-
-    #[napi]
-    pub fn get_inet(&self) -> napi::Result<InetAddressWrapper> {
-        match self.inner.as_inet() {
-            Some(r) => Ok(InetAddressWrapper::from_ip_addr(r)),
-            None => Err(Self::generic_error("inet")),
-        }
-    }
-
-    #[napi]
-    pub fn get_list(&self) -> napi::Result<Vec<CqlValueWrapper>> {
-        match self.inner.as_list() {
-            Some(r) => Ok(r
-                .iter()
-                .map(|f| CqlValueWrapper { inner: f.clone() })
-                .collect()),
-            None => Err(Self::generic_error("list")),
-        }
-    }
-
-    #[napi]
-    /// Return array of "tuples" keeping (key, value) for each element in the map
-    /// Currently NAPI doesn't implement ToNapiValue for tuples,
-    /// but the returned typed here will be the same as if we returned tuple.
-    pub fn get_map(&self) -> napi::Result<Vec<Vec<CqlValueWrapper>>> {
-        match self.inner.as_map() {
-            Some(r) => Ok(r
-                .iter()
-                .map(|f| {
-                    vec![
-                        CqlValueWrapper { inner: f.0.clone() },
-                        CqlValueWrapper { inner: f.1.clone() },
-                    ]
-                })
-                .collect()),
-            None => Err(Self::generic_error("map")),
-        }
-    }
-
-    #[napi]
-    pub fn get_set(&self) -> napi::Result<Vec<CqlValueWrapper>> {
-        match self.inner.as_set() {
-            Some(r) => Ok(r
-                .iter()
-                .map(|f| CqlValueWrapper { inner: f.clone() })
-                .collect()),
-            None => Err(Self::generic_error("set")),
-        }
-    }
-
-    #[napi]
-    pub fn get_small_int(&self) -> napi::Result<i16> {
-        match self.inner.as_smallint() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("small_int")),
-        }
-    }
-
-    #[napi]
-    pub fn get_tiny_int(&self) -> napi::Result<i8> {
-        match self.inner.as_tinyint() {
-            Some(r) => Ok(r),
-            None => Err(Self::generic_error("tiny_int")),
-        }
-    }
-    #[napi]
-    pub fn get_local_time(&self) -> napi::Result<LocalTimeWrapper> {
-        match self.inner.as_cql_time() {
-            Some(r) => Ok(LocalTimeWrapper::from_cql_time(r)),
-            None => Err(Self::generic_error("local_time")),
-        }
-    }
-
-    #[napi]
-    pub fn get_time_uuid(&self) -> napi::Result<TimeUuidWrapper> {
-        match self.inner.as_timeuuid() {
-            Some(r) => Ok(TimeUuidWrapper::from_cql_time_uuid(r)),
-            None => Err(Self::generic_error("time_uuid")),
-        }
-    }
-
-    #[napi]
-    pub fn get_tuple(&self) -> napi::Result<Vec<Option<CqlValueWrapper>>> {
-        match &self.inner {
-            CqlValue::Tuple(r) => Ok(r
-                .iter()
-                .map(|v| v.as_ref().map(|e| CqlValueWrapper { inner: e.clone() }))
-                .collect()),
-            _ => Err(Self::generic_error("tuple")),
-        }
-    }
-
-    #[napi]
-    pub fn get_uuid(&self) -> napi::Result<UuidWrapper> {
-        match self.inner.as_uuid() {
-            Some(r) => Ok(UuidWrapper::from_cql_uuid(r)),
-            None => Err(Self::generic_error("uuid")),
         }
     }
 }