Replaces current approach of returning CQL values from rust driver to JS.
Instead of calling NAPI twice:
    - first to get value type
    - second to get value itself
we convert the value when we return CQL wrapper.

This improves both code performance, by reducing native calls
and (hopefully) increases code clarity.

For ``select`` benchmark ~20/15% improvement (task-clock/time elapsed)
After:

```bash
sudo perf stat node ./benchmark/logic/select.js scylladb-javascript-driver 300000

 Performance counter stats for 'node ./benchmark/logic/select.js scylladb-javascript-driver 300000':

         39,031.27 msec task-clock                       #    0.672 CPUs utilized             
         1,213,807      context-switches                 #   31.098 K/sec                     
            27,146      cpu-migrations                   #  695.494 /sec                      
            13,186      page-faults                      #  337.832 /sec                      
   156,124,848,515      cycles                           #    4.000 GHz                       
    52,655,487,799      stalled-cycles-frontend          #   33.73% frontend cycles idle      
   183,755,495,921      instructions                     #    1.18  insn per cycle            
                                                  #    0.29  stalled cycles per insn   
    34,349,924,021      branches                         #  880.062 M/sec                     
       828,124,176      branch-misses                    #    2.41% of all branches           

      58.087041914 seconds time elapsed

      30.632502000 seconds user
       9.542099000 seconds sys
```

Before:

```bash
sudo perf stat node ./benchmark/logic/select.js scylladb-javascript-driver 300000

 Performance counter stats for 'node ./benchmark/logic/select.js scylladb-javascript-driver 300000':

         49,734.48 msec task-clock                       #    0.721 CPUs utilized             
         1,213,460      context-switches                 #   24.399 K/sec                     
            26,151      cpu-migrations                   #  525.812 /sec                      
            11,224      page-faults                      #  225.678 /sec                      
   204,333,765,051      cycles                           #    4.108 GHz                       
    57,481,147,481      stalled-cycles-frontend          #   28.13% frontend cycles idle      
   245,798,008,598      instructions                     #    1.20  insn per cycle            
                                                  #    0.23  stalled cycles per insn   
    45,616,064,580      branches                         #  917.192 M/sec                     
       876,843,759      branch-misses                    #    1.92% of all branches           

      68.992424536 seconds time elapsed

      41.575543000 seconds user
       9.419669000 seconds sys
```

For ``concurrent_select`` benchmark ~20/25% improvement (task-clock/time elapsed)

After:

```bash
sudo perf stat  ./benchmark/logic/concurrent_select.js scylladb-javascript-driver 800000

 Performance counter stats for ' ./benchmark/logic/concurrent_select.js scylladb-javascript-driver 800000':

        122,188.18 msec task-clock                       #    1.497 CPUs utilized             
         2,913,728      context-switches                 #   23.846 K/sec                     
            65,356      cpu-migrations                   #  534.880 /sec                      
            65,304      page-faults                      #  534.454 /sec                      
   501,089,264,243      cycles                           #    4.101 GHz                       
   116,760,887,470      stalled-cycles-frontend          #   23.30% frontend cycles idle      
   531,594,212,335      instructions                     #    1.06  insn per cycle            
                                                  #    0.22  stalled cycles per insn   
    99,554,837,139      branches                         #  814.766 M/sec                     
     2,015,453,685      branch-misses                    #    2.02% of all branches           

      81.617669943 seconds time elapsed

      98.907247000 seconds user
      25.066759000 seconds sys
```

Before:

```bash
sudo perf stat  ./benchmark/logic/concurrent_select.js scylladb-javascript-driver 800000

 Performance counter stats for ' ./benchmark/logic/concurrent_select.js scylladb-javascript-driver 800000':

        156,468.26 msec task-clock                       #    1.384 CPUs utilized             
         2,974,658      context-switches                 #   19.011 K/sec                     
            77,333      cpu-migrations                   #  494.241 /sec                      
            76,393      page-faults                      #  488.233 /sec                      
   659,739,880,134      cycles                           #    4.216 GHz                       
   127,827,841,049      stalled-cycles-frontend          #   19.38% frontend cycles idle      
   701,988,202,011      instructions                     #    1.06  insn per cycle            
                                                  #    0.18  stalled cycles per insn   
   130,604,458,737      branches                         #  834.703 M/sec                     
     2,179,959,479      branch-misses                    #    1.67% of all branches           

     113.037014315 seconds time elapsed

     132.587983000 seconds user
      25.839390000 seconds sys
```